requirements.txt
pandas>=2.0
numpy>=1.24
microsoft-qlib>=0.9.0
pytrader @ git+https://github.com/jeanboydev/pytrader-api.git
openai>=1.0
requests>=2.31
